refactor: :recycle: Resolved bug in change comments of the pr, and resolved comments
