+++ src/catalogs/roles/controllers/roles.controller.ts
@@ -1,8 +1,10 @@
+import { Body } from '@nestjs/common';
+
 import { apiMethods, EModules, Swagger } from '@/config';
 import { GuardSwagger } from '@/config/swagger/decorators/guard-swagger.decorator';
-import { RolesService } from '../services/roles.services';
-import { CreateRoleDto } from '../dto/create-roles.dto';
-import { Body } from '@nestjs/common';
+
+import { CreateRoleDto } from '../dto';
+import { RolesService } from '../services';
 
 @GuardSwagger({
   tag: 'roles',
