+++ src/catalogs/roles/roles.module.ts
@@ -2,15 +2,14 @@ import { Module } from '@nestjs/common';
 
 import { PrismaModule } from '@/config';
 
-import { RolesPrismaService } from './helpers';
 import { RolesController } from './controllers';
-import { AuthModule } from '@/providers';
+import { RolesPrismaService } from './helpers';
 import { RolesService } from './services/roles.services';
 
 @Module({
   providers: [RolesPrismaService, RolesService],
   controllers: [RolesController],
   exports: [RolesPrismaService],
-  imports: [PrismaModule, AuthModule],
+  imports: [PrismaModule],
 })
 export class RolesModule {}
