+++ src/modules/enterprises/enterprises.module.ts
@@ -2,11 +2,11 @@ import { Module } from '@nestjs/common';
 
 import { CitiesModule } from '@/catalogs';
 import { PrismaModule } from '@/config';
+import { FirebaseModule } from '@/providers/firebase';
 
 import { EnterprisesController } from './controllers';
 import { EnterprisesService } from './services';
 import { EnterprisesPrismaService } from './helpers/enterprises.prisma.helper.service';
-import { FirebaseModule } from '@/providers/firebase';
 
 @Module({
   controllers: [EnterprisesController],
